+++ pkg/tests/testthat/test-hybridModel.R
@@ -90,6 +90,13 @@ if(require(forecast) & require(testthat)){
     dat <- ts(rnorm(52 * 3), f = 52)
     expect_warning(hm <- hybridModel(y = dat, models = "fs"), NA)
     expect_true(length(forecast(hm)$mean) == 52 * 2)
+    # Test for messy function call
+    hm <- hybridModel(wineind)
+    # Less than orignal, messy function call
+    expect_true(format(object.size(hm)) < "6035456 bytes")
+    # No worse than improved function call
+    # Disable for now because fails on r-devel
+    #expect_true(format(object.size(hm)) <= "322352 bytes")
   })
 
   test_that("Testing model matching", {